moved RequestLog model to models.py, added admin view template
